(dstn) add an.util directory

+++ src/tilecache/an/tile/models.py
@@ -11,10 +11,10 @@ class Image(models.Model):
 	origformat = models.CharField(maxlength=30)
 	# The JPEG base filename.
 	filename = models.CharField(maxlength=1024)
-	ramin =	 models.DecimalField(decimal_places=7, max_digits=10)
-	ramax =	 models.DecimalField(decimal_places=7, max_digits=10)
-	decmin = models.DecimalField(decimal_places=7, max_digits=10)
-	decmax = models.DecimalField(decimal_places=7, max_digits=10)
+	ramin =	 models.FloatField()
+	ramax =	 models.FloatField()
+	decmin = models.FloatField()
+	decmax = models.FloatField()
 	imagew = models.IntegerField()
 	imageh = models.IntegerField()
